requirements.txt
streamlit
pandas
yfinance
plotly
pandas_ta
numpy
scikit-learn
statsmodels
python-dateutil
